+++ big-trader-api/src/main/java/com/flab/bigtrader/Fuck.java
@@ -1,0 +1,14 @@
+package com.flab.bigtrader;
+
+import javax.annotation.PostConstruct;
+
+import org.springframework.stereotype.Component;
+
+@Component
+public class Fuck {
+
+	@PostConstruct
+	public void dd() {
+		System.out.println("---------> 으악");
+	}
+}
