+++ big-trader-api/src/main/java/com/flab/bigtrader/apppush/application/AppPushAsyncBlockingService.java
@@ -6,7 +6,7 @@ import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 
 import com.flab.bigtrader.apppush.infrastructure.AppPushClient;
-import com.flab.bigtrader.apppush.infrastructure.AppPushSendEvent;
+import com.flab.bigtrader.apppush.infrastructure.dto.AppPushSendEvent;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -14,11 +14,11 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 @Service
 @RequiredArgsConstructor
-public class AppPushAsyncNonBlockingService implements AppPushService {
+public class AppPushAsyncBlockingService implements AppPushService {
 
-	private static final int TOTAL_LOOP_COUNT = 450_000;
+	private static final int TOTAL_LOOP_COUNT = 300_000;
 
-	private final AppPushClient nonBlockingAppPushClient;
+	private final AppPushClient blockingAppPushClient;
 
 	@Async
 	@Override
@@ -26,18 +26,15 @@ public class AppPushAsyncNonBlockingService implements AppPushService {
 		CountDownLatch countDownLatch = new CountDownLatch(TOTAL_LOOP_COUNT);
 
 		long start = System.currentTimeMillis();
-		log.info("시작시간: {}", start);
 
 		for (int i = 0; i < TOTAL_LOOP_COUNT; i++) {
-			nonBlockingAppPushClient.sendAppPush(appPushSendEvent, countDownLatch);
+			blockingAppPushClient.sendAppPush(appPushSendEvent, countDownLatch);
 		}
-
-		long end = System.currentTimeMillis();
 
 		awaitTotalLoopEnd(countDownLatch);
 
-		log.info("종료시간: {}", end);
-		log.info("총 시간: {}ms초", end - start);
+		long end = System.currentTimeMillis();
+		log.info("소요시간: {} 초", toSecond(start, end));
 	}
 
 	private void awaitTotalLoopEnd(CountDownLatch countDownLatch) {
@@ -46,5 +43,9 @@ public class AppPushAsyncNonBlockingService implements AppPushService {
 		} catch (InterruptedException e) {
 			log.info("전제 루프 종료");
 		}
+	}
+
+	private static long toSecond(long start, long end) {
+		return (end - start) / 1000;
 	}
 }
