+++ big-trader-api/src/main/java/com/flab/bigtrader/apppush/application/AppPushService.java
@@ -1,6 +1,6 @@
 package com.flab.bigtrader.apppush.application;
 
-import com.flab.bigtrader.apppush.infrastructure.AppPushSendEvent;
+import com.flab.bigtrader.apppush.infrastructure.dto.AppPushSendEvent;
 
 public interface AppPushService {
