+++ big-trader-api/src/main/java/com/flab/bigtrader/apppush/infrastructure/AppPushClient.java
@@ -2,6 +2,8 @@ package com.flab.bigtrader.apppush.infrastructure;
 
 import java.util.concurrent.CountDownLatch;
 
+import com.flab.bigtrader.apppush.infrastructure.dto.AppPushSendEvent;
+
 public interface AppPushClient {
 
 	void sendAppPush(AppPushSendEvent appPushSendEvent, CountDownLatch countDownLatch);
