+++ big-trader-api/src/main/java/com/flab/bigtrader/apppush/infrastructure/NonBlockingAppPushClient.java
@@ -8,27 +8,30 @@ import javax.annotation.PostConstruct;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
+import org.springframework.http.client.reactive.ReactorClientHttpConnector;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Component;
 import org.springframework.web.reactive.function.client.WebClient;
 
+import com.flab.bigtrader.apppush.infrastructure.dto.AppPushSendEvent;
 import com.flab.bigtrader.common.exception.AppPushServerConnectionException;
 import com.flab.bigtrader.common.exception.AppPushServerParameterException;
 
 import io.netty.handler.timeout.ReadTimeoutException;
 import lombok.extern.slf4j.Slf4j;
 import reactor.core.publisher.Mono;
+import reactor.netty.http.client.HttpClient;
+import reactor.netty.resources.ConnectionProvider;
 
 @Slf4j
 @Component
-public class BlockingAppPushClient implements AppPushClient {
+public class NonBlockingAppPushClient implements AppPushClient {
 
 	private final String appPushServerBaseUrl;
 
 	private WebClient webClient;
 
-	public BlockingAppPushClient(
+	public NonBlockingAppPushClient(
 		@Value("${app.push.base-url}")
 		String appPushServerBaseUrl) {
 		this.appPushServerBaseUrl = appPushServerBaseUrl;
@@ -36,15 +39,27 @@ public class BlockingAppPushClient implements AppPushClient {
 
 	@PostConstruct
 	private void setUpWebClient() {
+		ConnectionProvider connectionProvider = ConnectionProvider.builder("non-blocking-connection-provider")
+			.maxIdleTime(Duration.ofSeconds(20))
+			.evictInBackground(Duration.ofSeconds(30))
+			.pendingAcquireMaxCount(-1)
+			.maxIdleTime(Duration.ofSeconds(80))
+			.lifo()
+			.build();
+
+		ReactorClientHttpConnector clientHttpConnector = new ReactorClientHttpConnector(
+			HttpClient.create(connectionProvider));
+
 		this.webClient = WebClient.builder()
 			.baseUrl(this.appPushServerBaseUrl)
+			.clientConnector(clientHttpConnector)
 			.build();
 	}
 
 	@Async
 	@Override
 	public void sendAppPush(AppPushSendEvent appPushSendEvent, CountDownLatch countDownLatch) {
-		Mono<ResponseEntity<Void>> responseEntityMono = webClient.post()
+		webClient.post()
 			.uri("/api/v1/app-push")
 			.contentType(MediaType.APPLICATION_JSON)
 			.bodyValue(appPushSendEvent)
@@ -53,9 +68,12 @@ public class BlockingAppPushClient implements AppPushClient {
 			.onStatus(HttpStatus::is5xxServerError, clientResponse -> Mono.error(AppPushServerConnectionException::new))
 			.toBodilessEntity()
 			.timeout(Duration.ofSeconds(80))
-			.onErrorMap(ReadTimeoutException.class, AppPushServerConnectionException::new);
+			.onErrorMap(ReadTimeoutException.class, AppPushServerConnectionException::new)
+			.subscribe(result -> progressSuccess(countDownLatch), e -> log.error("실패", e.getCause()));
+	}
 
-		responseEntityMono.block();
+	private static void progressSuccess(CountDownLatch countDownLatch) {
 		countDownLatch.countDown();
+		log.info("성공");
 	}
 }
