+++ big-trader-api/src/main/java/com/flab/bigtrader/common/kafka/serializer/CustomObjectSerializer.java
