--- big-trader-api/src/main/kotlin/com/flab/bigtrader/apppush/infrastructure/AppPushSendDto.kt
@@ -1,5 +1,0 @@
-package com.flab.bigtrader.apppush.infrastructure
-
-data class AppPushSendEvent(
-        val message: String
-)
