+++ big-trader-api/src/main/kotlin/com/flab/bigtrader/apppush/infrastructure/dto/AppPushSendDto.kt
@@ -1,0 +1,5 @@
+package com.flab.bigtrader.apppush.infrastructure.dto
+
+data class AppPushSendEvent(
+    val message: String
+)
