+++ big-trader-api/src/main/kotlin/com/flab/bigtrader/apppush/presentation/dto/AppPushDto.kt
@@ -1,9 +1,9 @@
 package com.flab.bigtrader.apppush.presentation.dto
 
-import com.flab.bigtrader.apppush.infrastructure.AppPushSendEvent
+import com.flab.bigtrader.apppush.infrastructure.dto.AppPushSendEvent
 
 data class AppPushRequest(
-        val message: String
+    val message: String
 ) {
     fun toAppPushSendEvent() = AppPushSendEvent(message = this.message)
 }
