+++ big-trader-api/src/test/java/com/flab/bigtrader/common/AcceptanceTest.java
