+++ config/redis-config/src/main/java/com/flab/bigtrader/config/RedisConfig.java
@@ -1,0 +1,8 @@
+package com.flab.bigtrader.config;
+
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class RedisConfig {
+
+}
